--- RollenBollen/templates/home.html
@@ -1,121 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <title>RollenBollen | QRM4.0</title>
-  <link rel="shortcut icon" href="https://external-content.duckduckgo.com/ip3/edu.sphero.com.ico" />
-  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-    integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
-    crossorigin="anonymous"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js"
-    integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
-    crossorigin="anonymous"></script>
-  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
-    integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous" />
-  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"
-    integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy"
-    crossorigin="anonymous"></script>
-  <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-  <style>
-    .flexxer {
-      display: flex;
-      margin: 5px 0px;
-    }
-
-    .flexxer>button {
-      margin: 0px 5px
-    }
-  </style>
-</head>
-
-<body>
-  <div class="container">
-    <div class="row mt-3 mb-5">
-      <div class="col">
-        <div class="row">
-          <div class="col">
-            <h1>Visualise algorithms</h1>
-            <div class="row">
-              <div class="col flexxer">
-                <button type="button" class="btn btn-success" id="drawPathDFS">
-                  Depth-First Search
-                </button>
-                <button type="button" class="btn btn-info" id="drawPathBFS">
-                  Breadth-First Search
-                </button>
-                <button type="button" class="btn btn-warning" id="drawPathAstar">
-                  A* Search
-                </button>
-              </div>
-            </div>
-            <h1>Visualise paths</h1>
-            <div class="row">
-              <div class="col flexxer">
-                <button type="button" class="btn btn-success" id="finalDFS">
-                  Depth-First Path
-                </button>
-                <button type="button" class="btn btn-info" id="finalBFS">
-                  Breadth-First Path
-                </button>
-              </div>
-            </div>
-            <div class="row">
-              <div class="col flexxer">
-                <button type="button" class="btn btn-warning" id="finalAstar">
-                  A* Search
-                </button>
-                <button type="button" class="btn btn-danger" id="optFinalAstar">
-                  Optimal A* Search
-                </button>
-              </div>
-            </div>
-            <h1>Generate a new random maze:</h1>
-            <div class="row">
-              <div class="col flexxer">
-                <form action="{{ url_for('page_home') }}" method="post">
-                  <button type="submit" class="btn btn-success" name="rand">
-                    Randomize Walls
-                  </button>
-                  <button type="submit" class="btn btn-info" name="rand" value="loc">
-                    Randomize Finish Location
-                  </button>
-                </form>
-              </div>
-            </div>
-            <h1>Server settings</h1>
-            <div class="row">
-              <div class="col flexxer">
-                <button class="btn btn-warning"
-                  onclick="const b = async ()=>{await fetch('{{ url_for('api_go_home') }}')};b()">
-                  Home
-                </button>
-                <button class="btn btn-danger"
-                  onclick="const a = async ()=>{await fetch('{{ url_for('reset_webserver') }}')};a()">
-                  Reset
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div class="col">
-        <div class="row">
-          <div class="col">
-            <table id="game"></table>
-            <script src="{{ url_for('static', filename='board/jsboard.min.js') }}"></script>
-            <script>
-              var maze_map = {{ maze_map| tojson }}; var dfs_path = {{ dfs_path| tojson }};
-              var bfs_path = {{ bfs_path| tojson }}; var astar_path = {{ astar_path| tojson }};
-              var final_dfs = {{ final_dfs| tojson }}; var final_bfs = {{ final_bfs| tojson }};
-              var final_astar = {{ final_astar| tojson }};
-              var optimal_final_astar = {{ optimal_final_astar| tojson }};
-            </script>
-            <script src="{{ url_for('static', filename='boardcreation.js') }}"></script>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</body>
-
-</html>
