+++ static/stats/rand_search_final.js
